backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
tensorflow-cpu 
pillow
numpy<2
python-dotenv